# docs/swagger.yaml
basePath: /
definitions:
  authapi.loginParams:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  sqlcategory.Category:
    properties:
      icon:
        type: string
      name:
        type: string
      subcategory:
        items:
          type: string
        type: array
    type: object
  sqlitem.CreateParams:
    properties:
      category:
        type: string
      condition:
        type: integer
      description:
        type: string
      images:
        items:
          type: string
        type: array
      is_negotiable:
        type: boolean
      name:
        type: string
      placeholder:
        type: string
      price:
        type: number
      subcategory:
        type: string
    type: object
  sqlitem.Item:
    properties:
      category:
        type: string
      condition:
        type: integer
      description:
        type: string
      fts:
        type: string
      id:
        type: integer
      images:
        items:
          type: string
        type: array
      is_negotiable:
        type: boolean
      name:
        type: string
      placeholder:
        type: string
      posted_at:
        type: string
      price:
        type: number
      sold_at:
        type: string
      subcategory:
        type: string
      views:
        type: integer
    type: object
  sqlitem.UpdateParams:
    properties:
      category:
        type: string
      condition:
        type: integer
      description:
        type: string
      id:
        type: integer
      images:
        items:
          type: string
        type: array
      is_negotiable:
        type: boolean
      name:
        type: string
      placeholder:
        type: string
      price:
        type: number
      subcategory:
        type: string
      views:
        type: integer
    type: object
  sqluser.CreateParams:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  sqluser.User:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  storage.PresignParams:
    properties:
      bucket:
        type: string
      contentType:
        type: string
      key:
        type: string
      method:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the REST API for DooleyOnline.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: DooleyOnline API
  version: "1.0"
paths:
  /:
    get:
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Greeting
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login Params
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/authapi.loginParams'
      produces:
      - application/json
      responses:
        "200":
          description: User
          schema:
            $ref: '#/definitions/sqluser.User'
      summary: Log in
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: Result
          schema:
            type: string
      summary: Log out
      tags:
      - auth
  /category:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/sqlcategory.Category'
            type: array
      summary: Get all categories
      tags:
      - category
  /category/{name}:
    get:
      parameters:
      - description: Category name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sqlcategory.Category'
      summary: Get category by name
      tags:
      - category
  /item:
    get:
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: Category filter
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/sqlitem.Item'
            type: array
      summary: Get many items
      tags:
      - item
    post:
      consumes:
      - application/json
      parameters:
      - description: Item
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/sqlitem.CreateParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sqlitem.Item'
      summary: Create item
      tags:
      - item
  /item/{id}:
    delete:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
      summary: Delete item by ID
      tags:
      - item
    get:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sqlitem.Item'
      summary: Get item by ID
      tags:
      - item
    put:
      consumes:
      - application/json
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Item
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/sqlitem.UpdateParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sqlitem.Item'
      summary: Update item by ID
      tags:
      - item
  /item/{id}/sell:
    post:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Item ID
          schema:
            type: string
      summary: Update sold_at property by ID
      tags:
      - item
  /item/{id}/view:
    post:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
      summary: Increment item views by ID
      tags:
      - item
  /storage/presign:
    post:
      consumes:
      - application/json
      parameters:
      - description: Presign params
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/storage.PresignParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sqlitem.Item'
      summary: Generate presigned URL for s3 operation
      tags:
      - other
  /user/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/sqluser.CreateParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sqluser.User'
      summary: Create user
      tags:
      - user
swagger: "2.0"